Test_backend/Test_backend/requirements.txt
pandas
fastapi
uvicorn
dotenv
pydantic
pymysql
SQLALchemy
pyJWT
alembic
fastapi-cors
bcrypt